cambios

+++ Controllers/admin/metodos.js
@@ -1,7 +1,16 @@
 const baseURL = "http://localhost:4000/api";
 const token = sessionStorage.getItem("token");
 
- async function fetchData(endpoint) {
+let isFetchingData = false;
+
+async function fetchData(endpoint) {
+    if (isFetchingData) {
+        console.log("Otra operación está en curso. Por favor, espere.");
+        return { success: false, message: "Otra operación está en curso. Por favor, espere." };
+    }
+
+    isFetchingData = true;
+
     try {
         const response = await fetch(`${baseURL}${endpoint}`, {
             headers: {
@@ -13,10 +22,19 @@ const token = sessionStorage.getItem("token");
     } catch (error) {
         console.error("Error:", error);
         return { success: false, message: error.message };
+    } finally {
+        isFetchingData = false;
     }
 }
 
- async function createData(endpoint, formData) {
+async function createData(endpoint, formData) {
+    if (isFetchingData) {
+        console.log("Otra operación está en curso. Por favor, espere.");
+        return { success: false, message: "Otra operación está en curso. Por favor, espere." };
+    }
+
+    isFetchingData = true;
+
     try {
         const response = await fetch(`${baseURL}${endpoint}`, {
             method: 'POST',
@@ -30,10 +48,19 @@ const token = sessionStorage.getItem("token");
     } catch (error) {
         console.error("Error:", error);
         return { success: false, message: error.message };
+    } finally {
+        isFetchingData = false;
     }
 }
 
- async function updateData(endpoint, formData) {
+async function updateData(endpoint, formData) {
+    if (isFetchingData) {
+        console.log("Otra operación está en curso. Por favor, espere.");
+        return { success: false, message: "Otra operación está en curso. Por favor, espere." };
+    }
+
+    isFetchingData = true;
+
     try {
         const response = await fetch(`${baseURL}${endpoint}`, {
             method: 'PUT',
@@ -47,10 +74,19 @@ const token = sessionStorage.getItem("token");
     } catch (error) {
         console.error("Error:", error);
         return { success: false, message: error.message };
+    } finally {
+        isFetchingData = false;
     }
 }
 
- async function deleteData(endpoint) {
+async function deleteData(endpoint) {
+    if (isFetchingData) {
+        console.log("Otra operación está en curso. Por favor, espere.");
+        return { success: false, message: "Otra operación está en curso. Por favor, espere." };
+    }
+
+    isFetchingData = true;
+
     try {
         const response = await fetch(`${baseURL}${endpoint}`, {
             method: 'DELETE',
@@ -63,5 +99,7 @@ const token = sessionStorage.getItem("token");
     } catch (error) {
         console.error("Error:", error);
         return { success: false, message: error.message };
+    } finally {
+        isFetchingData = false;
     }
 }
